+++ Sources/ViewModels/ALKConversationViewModel.swift
@@ -104,7 +104,15 @@ open class ALKConversationViewModel: NSObject, Localizable {
         }
         return alchannel.type == 6
     }
-
+    
+    // To get Conversation created time based on its first message.
+    open var conversationCreatedTime: NSNumber? {
+        if alMessages.isEmpty {
+            return nil
+        }
+        return alMessages[0].createdAtTime
+    }
+    
     private var conversationId: NSNumber? {
         return conversationProxy?.id
     }
@@ -128,9 +136,7 @@ open class ALKConversationViewModel: NSObject, Localizable {
     private var awsEncryptionPrefix = "AWS-ENCRYPTED"
     private var botDelayTime = 0
 
-    
     // MARK: - Initializer
-
     public required init(
         contactId: String?,
         channelKey: NSNumber?,
@@ -1203,6 +1209,10 @@ open class ALKConversationViewModel: NSObject, Localizable {
         conversationProfile.status = ALKConversationProfile.Status(isOnline: contact.connected, lastSeenAt: contact.lastSeenAt)
         return conversationProfile
     }
+    
+    open func checkForTextToSpeech(list: [ALMessage]) {
+        KMTextToSpeech.shared.addMessagesToSpeech(list)
+    }
    
     func loadMessages() {
         var time: NSNumber?
@@ -1224,14 +1234,16 @@ open class ALKConversationViewModel: NSObject, Localizable {
 
             self.alMessages = messages.reversed() as! [ALMessage]
             self.alMessageWrapper.addObject(toMessageArray: messages)
+            
             self.modelsToBeAddedAfterDelay = self.alMessages.map { $0.messageModel }
             // Check for Conversation Assignee and conversation first message created time to show Typing Indicator.
             if self.isConversationAssignedToBot() && (self.botDelayTime > 0) && !self.isOldConversation() {
                 self.showTypingIndicatorForWelcomeMessage()
             } else {
                 self.messageModels = self.modelsToBeAddedAfterDelay
+               
             }
-
+            
             let showLoadEarlierOption: Bool = self.messageModels.count >= 50
             ALUserDefaultsHandler.setShowLoadEarlierOption(showLoadEarlierOption, forContactId: self.chatId)
             self.membersInGroup { members in
@@ -1269,7 +1281,7 @@ open class ALKConversationViewModel: NSObject, Localizable {
     // Check for Old Conversation based on created time
     func isOldConversation() -> Bool {
         let createdTimeInMilliSec = self.alMessages[0].createdAtTime as? Double ?? 0.0
-        let date = NSDate() 
+        let date = NSDate()
         let currentTimeInMilliSec = date.timeIntervalSince1970 * 1000
         let diff = currentTimeInMilliSec - createdTimeInMilliSec
         // Checking time difference of 10 seconds.